app_reemplazos/py/requirements.txt
streamlit
streamlit-authenticator
pandas
pyyaml
pillow
